# Conflicts:
#	SConstruct
#	src/dragengine/SConscript
#	src/launcher/android/DELauncher/.idea/caches/build_file_checksums.ser
#	src/launcher/android/DELauncher/.idea/misc.xml
#	src/launcher/android/DELauncher/app/app.iml

+++ src/dragengine/src/filesystem/deVFSDiskDirectory.cpp
@@ -1,41 +1,45 @@
-/* 
- * Drag[en]gine Game Engine
+/*
+ * MIT License
  *
- * Copyright (C) 2020, Roland Plüss (roland@rptd.ch)
- * 
- * This program is free software; you can redistribute it and/or 
- * modify it under the terms of the GNU General Public License 
- * as published by the Free Software Foundation; either 
- * version 2 of the License, or (at your option) any later 
- * version.
+ * Copyright (C) 2024, DragonDreams GmbH (info@dragondreams.ch)
  *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
  */
 
 #include <fcntl.h>
 #include <stdlib.h>
 #include <string.h>
 
-#include <unistd.h>
-#include <dirent.h>
-
 #include "../dragengine_configuration.h"
-#ifdef OS_MACOS
+
+#ifndef OS_W32_VS
+#	include <unistd.h>
+#	include <dirent.h>
+#endif
+
+#if defined OS_MACOS
 #	include <sys/time.h>
 #	include <fnmatch.h>
 #	include <errno.h>
 #elif defined OS_UNIX
 #	include <errno.h>
 #	include <fnmatch.h>
-#elif defined OS_W32
-#	include "../app/include_windows.h"
 #endif
 #include <sys/types.h>
 #include <sys/stat.h>
@@ -92,6 +96,12 @@ pDiskPath( diskPath ),
 pReadOnly( false ){
 }
 
+deVFSDiskDirectory::deVFSDiskDirectory( const decPath &rootPath, const decPath &diskPath, bool readonly ) :
+deVFSContainer( rootPath ),
+pDiskPath( diskPath ),
+pReadOnly( readonly ){
+}
+
 deVFSDiskDirectory::~deVFSDiskDirectory(){
 }
 
@@ -108,7 +118,11 @@ bool deVFSDiskDirectory::ExistsFile( const decPath &path ){
 #ifdef OS_W32
 	wchar_t widePath[ MAX_PATH ];
 	deOSWindows::Utf8ToWide( ( pDiskPath + path ).GetPathNative(), widePath, MAX_PATH );
+	#ifdef OS_W32_VS
+	return _waccess_s( widePath, 0 ) == 0; // file exists
+	#else
 	return _waccess( widePath, F_OK ) == 0;
+	#endif
 	
 #else
 	return access( ( pDiskPath + path ).GetPathNative(), F_OK ) == 0;
@@ -119,7 +133,11 @@ bool deVFSDiskDirectory::CanReadFile( const decPath &path ){
 #ifdef OS_W32
 	wchar_t widePath[ MAX_PATH ];
 	deOSWindows::Utf8ToWide( ( pDiskPath + path ).GetPathNative(), widePath, MAX_PATH );
+	#ifdef OS_W32_VS
+	return _waccess_s( widePath, 4 ) == 0; // read only
+	#else
 	return _waccess( widePath, R_OK ) == 0;
+	#endif
 	
 #else
 	return access( ( pDiskPath + path ).GetPathNative(), R_OK ) == 0;
@@ -138,9 +156,15 @@ bool deVFSDiskDirectory::CanWriteFile( const decPath &path ){
 	wchar_t widePath[ MAX_PATH ];
 	
 	deOSWindows::Utf8ToWide( diskPath.GetPathNative(), widePath, MAX_PATH );
-	if( _waccess( widePath, F_OK ) == 0 ){
-		canWrite = ( _waccess( widePath, W_OK ) == 0 );
+	#ifdef OS_W32_VS
+	if( _waccess_s( widePath, 0 ) == 0 ){ // file exists
+		canWrite = _waccess_s( widePath, 2 ) == 0; // write-only allowed
 	}
+	#else
+	if( _waccess( widePath, F_OK ) == 0 ){
+		canWrite = _waccess( widePath, W_OK ) == 0;
+	}
+	#endif
 	
 	if( canWrite ){
 		while( diskPath.GetComponentCount() > 0 ){
@@ -149,7 +173,11 @@ bool deVFSDiskDirectory::CanWriteFile( const decPath &path ){
 			deOSWindows::Utf8ToWide( diskPath.GetPathNative(), widePath, MAX_PATH );
 			if( GetFileAttributesExW( widePath, GetFileExInfoStandard, &fa ) ){
 				if( ( fa.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY ) == FILE_ATTRIBUTE_DIRECTORY ){
+					#ifdef OS_W32_VS
+					canWrite = _waccess_s( widePath, 2 ) == 0; // write-only allowed
+					#else
 					canWrite = _waccess( widePath, W_OK ) == 0;
+					#endif
 					break;
 					
 				}else{
@@ -209,14 +237,23 @@ bool deVFSDiskDirectory::CanDeleteFile( const decPath &path ){
 	wchar_t widePath[ MAX_PATH ];
 	
 	deOSWindows::Utf8ToWide( diskPath.GetPathNative(), widePath, MAX_PATH );
+	#ifdef OS_W32_VS
+	if( _waccess_s( widePath, 0 ) == 0 ){ // file exists
+		canDelete = _waccess_s( widePath, 2 ) == 0; // write only
+	#else
 	if( _waccess( widePath, F_OK ) == 0 ){
 		canDelete = _waccess( widePath, W_OK ) == 0;
+	#endif
 		
 		if( canDelete ){
 			diskPath.RemoveLastComponent();
 			
 			deOSWindows::Utf8ToWide( diskPath.GetPathNative(), widePath, MAX_PATH );
+			#ifdef OS_W32_VS
+			canDelete = _waccess_s( widePath, 2 ) == 0; // write only
+			#else
 			canDelete = _waccess( widePath, W_OK ) == 0;
+			#endif
 		}
 	}
 	
@@ -259,13 +296,39 @@ void deVFSDiskDirectory::DeleteFile( const decPath &path ){
 #ifdef OS_W32
 	wchar_t widePath[ MAX_PATH ];
 	deOSWindows::Utf8ToWide( ( pDiskPath + path ).GetPathNative(), widePath, MAX_PATH );
-	if( _wunlink( widePath ) != 0 )
-#else
-	if( unlink( ( pDiskPath + path ).GetPathNative() ) != 0 )
-#endif
-	{
-		DETHROW_INFO( deeWriteFile, ( pDiskPath + path ).GetPathNative() );
+	
+	WIN32_FILE_ATTRIBUTE_DATA fa;
+	if( ! GetFileAttributesExW( widePath, GetFileExInfoStandard, &fa ) ){
+		DETHROW_INFO( deeFileNotFound, ( pDiskPath + path ).GetPathNative() );
 	}
+	
+	if( ( fa.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY ) == FILE_ATTRIBUTE_DIRECTORY ){
+		if( _wrmdir( widePath ) != 0 ){
+			DETHROW_INFO( deeWriteFile, ( pDiskPath + path ).GetPathNative() );
+		}
+		
+	}else{
+		if( _wunlink( widePath ) != 0 ){
+			DETHROW_INFO( deeWriteFile, ( pDiskPath + path ).GetPathNative() );
+		}
+	}
+#else
+	struct stat st;
+	if( stat( ( pDiskPath + path ).GetPathNative(), &st ) ){
+		DETHROW_INFO( deeFileNotFound, ( pDiskPath + path ).GetPathNative() );
+	}
+	
+	if( S_ISDIR( st.st_mode ) ){
+		if( rmdir( ( pDiskPath + path ).GetPathNative() ) != 0 ){
+			DETHROW_INFO( deeWriteFile, ( pDiskPath + path ).GetPathNative() );
+		}
+		
+	}else{
+		if( unlink( ( pDiskPath + path ).GetPathNative() ) != 0 ){
+			DETHROW_INFO( deeWriteFile, ( pDiskPath + path ).GetPathNative() );
+		}
+	}
+#endif
 }
 
 void deVFSDiskDirectory::TouchFile( const decPath &path ){
@@ -279,7 +342,11 @@ void deVFSDiskDirectory::TouchFile( const decPath &path ){
 #ifdef OS_W32
 	wchar_t widePath[ MAX_PATH ];
 	deOSWindows::Utf8ToWide( npath, widePath, MAX_PATH );
+	#ifdef OS_W32_VS
+	if( _waccess_s( widePath, 0 ) == 0 ){ // file exists
+	#else
 	if( _waccess( widePath, F_OK ) == 0 ){
+	#endif
 		HANDLE hfile = CreateFileW( widePath, FILE_WRITE_ATTRIBUTES, 0, NULL,
 			OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL, NULL );
 		SYSTEMTIME systime;
@@ -345,6 +412,38 @@ void deVFSDiskDirectory::SearchFiles( const decPath &directory, deContainerFileS
 				continue;
 			}
 			
+			#ifdef OS_BEOS
+			// missing d_type and DT_* in dirent
+			decPath pathLink( searchPath );
+			pathLink.AddComponent( entry->d_name );
+			const decString strPathLink( pathLink.GetPathNative() );
+			
+			struct stat st;
+			lstat( strPathLink, &st );
+			
+			if( S_ISREG( st.st_mode ) ){
+				searcher.Add( entry->d_name, deVFSContainer::eftRegularFile );
+				
+			}else if( S_ISDIR( st.st_mode ) ){
+				searcher.Add( entry->d_name, deVFSContainer::eftDirectory );
+				
+			}else if( S_ISLNK( st.st_mode ) ){
+				if( stat( strPathLink, &st ) ){
+					// dangling link. assume it is a file
+					searcher.Add( entry->d_name, deVFSContainer::eftRegularFile );
+					
+				}else if( S_ISREG( st.st_mode ) ){
+					searcher.Add( entry->d_name, deVFSContainer::eftRegularFile );
+					
+				}else if( S_ISDIR( st.st_mode ) ){
+					searcher.Add( entry->d_name, deVFSContainer::eftDirectory );
+					
+				}else{
+					searcher.Add( entry->d_name, deVFSContainer::eftSpecial );
+				}
+			}
+			
+			#else
 			if( entry->d_type == DT_REG ){
 				searcher.Add( entry->d_name, deVFSContainer::eftRegularFile );
 				
@@ -370,6 +469,7 @@ void deVFSDiskDirectory::SearchFiles( const decPath &directory, deContainerFileS
 					searcher.Add( entry->d_name, deVFSContainer::eftSpecial );
 				}
 			}
+			#endif
 		}
 		
 		closedir( theDir );
@@ -383,12 +483,15 @@ void deVFSDiskDirectory::SearchFiles( const decPath &directory, deContainerFileS
 	
 #elif defined OS_W32
 	HANDLE searchHandle = INVALID_HANDLE_VALUE;
-	wchar_t widePath[ MAX_PATH ];
+	wchar_t widePath[ MAX_PATH ], widePathSymlink[ MAX_PATH ];
 	WIN32_FIND_DATAW dirEntry;
+	BY_HANDLE_FILE_INFORMATION symlinkInfo;
 	DWORD lastError;
 	
 	searchPath.AddComponent( "*" );
 	deOSWindows::Utf8ToWide( searchPath.GetPathNative(), widePath, MAX_PATH );
+	
+	const size_t offsetWidePathSymlink = wcslen( widePath ) - 1;
 	
 	try{
 		searchHandle = FindFirstFileW( widePath, &dirEntry );
@@ -400,10 +503,29 @@ void deVFSDiskDirectory::SearchFiles( const decPath &directory, deContainerFileS
 			
 		}else{
 			while( true ){
+				if( ( dirEntry.dwFileAttributes & FILE_ATTRIBUTE_REPARSE_POINT ) == FILE_ATTRIBUTE_REPARSE_POINT ){
+					// consider invalid (a file) until proven otherwise
+					dirEntry.dwFileAttributes = FILE_ATTRIBUTE_NORMAL;
+
+					wcscpy_s( widePathSymlink, MAX_PATH, widePath );
+					wcscpy_s( widePathSymlink + offsetWidePathSymlink,
+						MAX_PATH - offsetWidePathSymlink, dirEntry.cFileName );
+
+					const HANDLE sfh = CreateFileW( widePathSymlink, 0, 0,
+						nullptr, OPEN_EXISTING, FILE_FLAG_BACKUP_SEMANTICS, nullptr );
+
+					if( sfh != INVALID_HANDLE_VALUE ){
+						if( GetFileInformationByHandle( sfh, &symlinkInfo ) ){
+							dirEntry.dwFileAttributes = symlinkInfo.dwFileAttributes;
+						}
+						CloseHandle( sfh );
+					}
+				}
+
 				const decString entryName( deOSWindows::WideToUtf8( dirEntry.cFileName ) );
 				
 				if( entryName != "." && entryName != ".." ){
-					if( ( dirEntry.dwFileAttributes | FILE_ATTRIBUTE_DIRECTORY ) == FILE_ATTRIBUTE_DIRECTORY ){
+					if( ( dirEntry.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY ) == FILE_ATTRIBUTE_DIRECTORY ){
 						searcher.Add( entryName, deVFSContainer::eftDirectory );
 						
 					}else{
